+++ relay-server/src/utils/rate_limits.rs
@@ -105,6 +105,7 @@ fn infer_event_category(item: &Item) -> Option<DataCategory> {
         ItemType::UserReport => None,
         ItemType::Profile => None,
         ItemType::ClientReport => None,
+        ItemType::Unknown(_) => None,
     }
 }
 
@@ -141,7 +142,7 @@ impl EnvelopeSummary {
         for item in envelope.items() {
             if item.creates_event() {
                 summary.infer_category(item);
-            } else if item.ty() == ItemType::Attachment {
+            } else if item.ty() == &ItemType::Attachment {
                 // Plain attachments do not create events.
                 summary.has_plain_attachments = true;
             }
@@ -411,7 +412,7 @@ where
         }
 
         // Remove attachments, except those required for processing
-        if enforcement.attachments.is_active() && item.ty() == ItemType::Attachment {
+        if enforcement.attachments.is_active() && item.ty() == &ItemType::Attachment {
             if item.creates_event() {
                 item.set_rate_limited(true);
                 return true;
@@ -421,7 +422,7 @@ where
         }
 
         // Remove sessions independently of events
-        if enforcement.sessions.is_active() && item.ty() == ItemType::Session {
+        if enforcement.sessions.is_active() && item.ty() == &ItemType::Session {
             return false;
         }